Improve Baysian deploy 111822

+++ utils/Baysian_deploy.py
@@ -40,33 +40,50 @@ def BayesianGranularityDeploy(file,path,kernel_size,trace_a,trace_b,thold,pathOu
     img = AIPS_pose_object.cellpose_image_load()
     # create mask for the entire image
     mask, table = AIPS_pose_object.cellpose_segmantation(image_input=img)
-    gran = ag.GRANULARITY(image=img, mask=mask)
-    granData = gran.loopLabelimage(start_kernel=1, end_karnel=kernel_size, kernel_size=kernel_size)
-    granDataFinal = ag.MERGE().calcDecay(granData, kernel_size)
-    def classify(n, thold):
-        mu = trace_a + trace_b * n
-        prob = 1 / (1 + np.exp(-mu))
-        return prob, prob > thold
-    rate = granDataFinal.intensity.values
-    prob, prediction = classify(rate, thold)
-    table["predict"] = prob
-    image_blank = np.zeros_like(img)
-    binary, table_sel = AIPS_pose_object.call_bin(table_sel_cor=table, threshold=0.9, img_blank=image_blank)
-    img_gs = img_as_ubyte(binary)
-    if saveMerge:
-        table['predict'] = np.round(table.predict.values, 2)
-        maskKeep = AIPS_pose_object.keepObject(table = table_sel)
-        compsiteImage = afd.Compsite_display(input_image=img, mask_roi=maskKeep)
-        LabeldImage = compsiteImage.display_image_label(table=table, font_select="arial.ttf", font_size=14,label_draw='predict', intensity=1)
-        LabeldImage.save(os.path.join(pathOut, id_generator() + '.png'))
-    if os.path.exists(os.path.join(pathOut, 'binary.tif')):
-        os.remove(os.path.join(pathOut, 'binary.tif'))
-    tfi.imsave(os.path.join(pathOut, 'binary.tif'), img_gs)
-    with open(os.path.join(pathOut, 'cell_count.txt'), 'r') as f:
-        prev_number = f.readlines()
-    new_value = int(prev_number[0]) + len(table_sel)
-    with open(os.path.join(pathOut, 'cell_count.txt'), 'w') as f:
-        f.write(str(new_value))
+    if len(table)<5:
+        with open(os.path.join(pathOut, 'cell_count.txt'), 'r') as f:
+            prev_number = f.readlines()
+        new_value = int(prev_number[0]) + len(table)
+        with open(os.path.join(pathOut, 'cell_count.txt'), 'w') as f:
+            f.write(str(new_value))
+        with open(os.path.join(pathOut, 'count.txt'), 'w') as f:
+            f.write(str(len(table)))
+    else:
+        gran = ag.GRANULARITY(image=img, mask=mask)
+        granData = gran.loopLabelimage(start_kernel=1, end_karnel=kernel_size, kernel_size=kernel_size)
+        granDataFinal = ag.MERGE().calcDecay(granData, kernel_size)
+        def classify(n, thold):
+            mu = trace_a + trace_b * n
+            prob = 1 / (1 + np.exp(-mu))
+            return prob, prob > thold
+        rate = granDataFinal.intensity.values
+        prob, prediction = classify(rate, thold)
+        table["predict"] = prob
+        image_blank = np.zeros_like(img)
+        binary, table_sel = AIPS_pose_object.call_bin(table_sel_cor=table, threshold=0.9, img_blank=image_blank)
+        img_gs = img_as_ubyte(binary)
+        if saveMerge:
+            table['predict'] = np.round(table.predict.values, 2)
+            maskKeep = AIPS_pose_object.keepObject(table = table_sel)
+            compsiteImage = afd.Compsite_display(input_image=img, mask_roi=maskKeep)
+            LabeldImage = compsiteImage.display_image_label(table=table, font_select="arial.ttf", font_size=14,label_draw='predict', intensity=1)
+            LabeldImage.save(os.path.join(pathOut, id_generator() + '.png'))
+        with open(os.path.join(pathOut, 'active_cell_count.txt'), 'r') as f:
+            prev_number_active = f.readlines()
+        new_value = int(prev_number_active[0]) + len(table_sel)
+        with open(os.path.join(pathOut, 'active_cell_count.txt'), 'w') as f:
+            f.write(str(new_value))
+        if os.path.exists(os.path.join(pathOut, 'binary.tif')):
+            os.remove(os.path.join(pathOut, 'binary.tif'))
+        tfi.imsave(os.path.join(pathOut, 'binary.tif'), img_gs)
+        with open(os.path.join(pathOut, 'cell_count.txt'), 'r') as f:
+            prev_number = f.readlines()
+        new_value = int(prev_number[0]) + len(rate)
+        with open(os.path.join(pathOut, 'cell_count.txt'), 'w') as f:
+            f.write(str(new_value))
+        with open(os.path.join(pathOut, 'count.txt'), 'w') as f:
+            f.write(str(len(table)))
+
 
 
 def BayesianGranularityDeployTest(file,path,kernel_size,trace_a,trace_b,thold,pathOut, clean ):
@@ -103,7 +120,6 @@ def BayesianGranularityDeployTest(file,path,kernel_size,trace_a,trace_b,thold,pa
     image_blank = np.zeros_like(img)
     binary, table_sel = AIPS_pose_object.call_bin(table_sel_cor=table, threshold=0.9, img_blank=image_blank)
     return img, mask, table, binary, table_sel
-
